pyproject.toml
[project]
name = "BasicTemplate"
version = "0.1.0"
description = "A Basic Template for any api"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "cryptography>=45.0.7",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.9",
    "pyjwt>=2.10.1",
    "redis>=6.4.0",
    "sqlalchemy[asyncio]>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]
